+++ bigml/api.py
@@ -1497,6 +1497,32 @@ class BigML(object):
         if dataset_id:
             return self._delete("%s%s" % (self.url, dataset_id))
 
+    def error_counts(self, dataset, raise_on_error=True):
+        """Returns the ids of the fields that contain errors and their number.
+
+           The dataset argument can be either a dataset resource structure
+           or a dataset id (that will be used to retrieve the associated
+           remote resource).
+
+        """
+        errors_dict = {}
+        if not isinstance(dataset, dict) or not 'object' in dataset: 
+            check_resource_type(dataset, DATASET_PATH,
+                                message="A dataset id is needed.")
+            dataset_id = get_dataset_id(dataset)
+            dataset = check_resource(dataset_id, self.get_dataset,
+                                     raise_on_error=raise_on_error)
+            if not raise_on_error and resource['error'] is not None:
+                dataset_id = None
+        else:
+            dataset_id = get_dataset_id(dataset)
+        if dataset_id:
+            errors = dataset.get('object', {}).get(
+                'status', {}).get('field_errors', {})
+            for field_id in errors:
+                errors_dict[field_id] = errors[field_id]['total']
+        return errors_dict
+
     ##########################################################################
     #
     # Models