+++ tests/features/terrain.py
@@ -53,7 +53,7 @@ def setup_resources(feature):
     world.batch_predictions = []
 
     world.dataset_ids = []
-
+    world.fields_properties_dict = {}
 @after.each_feature
 def cleanup_resources(feature):